Add a default shell explicitly.
Previously, the default shell from the underlying container was used
(according to the useradd man page), but that wasn't set to anything,
resulting in no shell being set in /etc/passwd.

That has been fine for everything so far, but the recent Globus
Connect Personal addition in dockerfiles/globus/ requires users to be
able to log in to FTP server contained within; and that needs the
shell to be a valid shell.

/bin/sh will hopefully be OK.