Add a username-specific tag to each built docker image
Previously my intention was that docker images would be anonymous
with respect to the tagging system: a docker image ID would be a
function of the environment Dockerfile, and `docker build` would be
run each time to get the appropriate docker image ID. The build cache
would mean that this didn't actually build the image every time.

However distinct from this is how to garbage collect old images: there
is no good story in docker that I know for this. With just a pile of
anonymous images in the list `docker images` output list, it was hard
to know which images were old, and which were your vital local
environments that your run every day.

This commit adds a tag to each built docker environment, of the form
`cue/$USERNAME/$ENVIRONMENT` which means that "docker images" will now
show images that belong to cue.

This introduces some shared namespace that did not previously exist,
though: multiple versions of cue might have different environments
with the same name.

The pragmatic decision I mdae is to disambiguate those by username, so
that users do not interfere with each other. Within a single user account,
it is up to that user to ensure that things do not interfere with each
other.