Implement alias mode if invoked not as `cue`
Alias mode will turn off the normal commandline
option processing.

Instead, it will run an environment with the same name as
the invoked command, and inside that environment again
invoke the same named command, passing all command line
options through to that command.

This is intended for use with symbolic links giving aliases
to cue that name command-specific environments.

Two environments are reconfigured to work that way in this
patch: hub and pijul